+++ java-client/src/test/java/org/opensearch/client/opensearch/_types/query_dsl/NeuralQueryTest.java
@@ -10,10 +10,11 @@ package org.opensearch.client.opensearch._types.query_dsl;
 
 import org.junit.Test;
 import org.opensearch.client.opensearch.model.ModelTestCase;
+import org.opensearch.client.util.MissingRequiredPropertiesException;
 
 public class NeuralQueryTest extends ModelTestCase {
     @Test
-    public void toBuilder() {
+    public void toBuilder_queryText() {
         NeuralQuery origin = new NeuralQuery.Builder().field("field")
             .queryText("queryText")
             .k(1)
@@ -22,5 +23,38 @@ public class NeuralQueryTest extends ModelTestCase {
         NeuralQuery copied = origin.toBuilder().build();
 
         assertEquals(toJson(copied), toJson(origin));
+    }
+
+    @Test
+    public void toBuilder_queryImage() {
+        NeuralQuery origin = new NeuralQuery.Builder().field("field")
+            .queryImage("queryImage")
+            .k(1)
+            .filter(IdsQuery.of(builder -> builder.values("Some_ID")).toQuery())
+            .build();
+        NeuralQuery copied = origin.toBuilder().build();
+
+        assertEquals(toJson(copied), toJson(origin));
+    }
+
+    @Test
+    public void toBuilder_both() {
+        NeuralQuery origin = new NeuralQuery.Builder().field("field")
+            .queryText("queryText")
+            .queryImage("queryImage")
+            .k(1)
+            .filter(IdsQuery.of(builder -> builder.values("Some_ID")).toQuery())
+            .build();
+        NeuralQuery copied = origin.toBuilder().build();
+
+        assertEquals(toJson(copied), toJson(origin));
+    }
+
+    @Test
+    public void toBuilder_missing_query() {
+        assertThrows(
+            MissingRequiredPropertiesException.class,
+            () -> new NeuralQuery.Builder().field("field").k(1).filter(IdsQuery.of(builder -> builder.values("Some_ID")).toQuery()).build()
+        );
     }
 }